Project compiles now with --std=c++11 flag.
Change the used version of c++ to c++11. This makes it possible
to use some c++ features which were not previously available.
To complie the project with the new c++ version, I had to change
some static variables in the basic movements and ransac class.
If this causes some trouble compiling the project, try to redo
this commit.